Transactional API for the Trie database back-end
This incorporates code from Nimbus's CachingDB, but also takes
into account the required ref counting needed by the current
Trie implementation and adds support for multiple outstanding
nested transactions.

A transaction is started with `beginTransaction`, followed by
a call to either `commit` or `rollback`.

`dispose` is a convenient destuctor-like proc that can be safely
executed in a `finally` block or a `defer` statement. It will
rollback the transaction unless it was already commited.